te mporary

+++ src/application/haptic/mp_t_haptic.cc
@@ -34,11 +34,11 @@ haptic::haptic(lib::configurator &_config) :
 void haptic::configure_edp_force_sensor(bool configure_track, bool configure_postument)
 {
 	if (configure_track) {
-		set_next_ecps_state((int) ecp_mp::task::ECP_GEN_BIAS_EDP_FORCE, 0, "", 0, 1, lib::ROBOT_IRP6OT_M.c_str());
+		set_next_ecps_state(ecp_mp::task::ECP_GEN_BIAS_EDP_FORCE, 0, "", 0, 1, lib::ROBOT_IRP6OT_M.c_str());
 	}
 
 	if (configure_postument) {
-		set_next_ecps_state((int) ecp_mp::task::ECP_GEN_BIAS_EDP_FORCE, 0, "", 0, 1, lib::ROBOT_IRP6P_M.c_str());
+		set_next_ecps_state(ecp_mp::task::ECP_GEN_BIAS_EDP_FORCE, 0, "", 0, 1, lib::ROBOT_IRP6P_M.c_str());
 	}
 
 	if ((configure_track) && (!configure_postument)) {
@@ -61,8 +61,8 @@ void haptic::main_task_algorithm(void)
 	configure_edp_force_sensor(true, true);
 
 	// wlaczenie generatora transparentnego w obu robotach
-	set_next_ecps_state((int) ecp_mp::task::ECP_GEN_TRANSPARENT, (int) 0, "", 0, 1, lib::ROBOT_IRP6OT_M.c_str());
-	set_next_ecps_state((int) ecp_mp::task::ECP_GEN_TRANSPARENT, (int) 0, "", 0, 1, lib::ROBOT_IRP6P_M.c_str());
+	set_next_ecps_state(ecp_mp::task::ECP_GEN_TRANSPARENT, (int) 0, "", 0, 1, lib::ROBOT_IRP6OT_M.c_str());
+	set_next_ecps_state(ecp_mp::task::ECP_GEN_TRANSPARENT, (int) 0, "", 0, 1, lib::ROBOT_IRP6P_M.c_str());
 
 	// mp_h_gen.sensor_m = sensor_m;
 	mp_h_gen.configure(1, 0);
